add feetlicker, update instructions, move files around to avoid project conflict
