+++ dotnet/libs-comparison.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "libs-comparison", "src\libs-comparison.csproj", "{BDFBF9B4-64C9-45A5-A996-854FB4314FA2}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{BDFBF9B4-64C9-45A5-A996-854FB4314FA2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BDFBF9B4-64C9-45A5-A996-854FB4314FA2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BDFBF9B4-64C9-45A5-A996-854FB4314FA2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BDFBF9B4-64C9-45A5-A996-854FB4314FA2}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
