+++ dotnet/src/Program.cs
@@ -1,0 +1,4 @@
+﻿using BenchmarkDotNet.Running;
+using libs_comparison;
+
+BenchmarkRunner.Run<ComparisonBenchmarks>();
