+++ dotnet/src/TwitchLib.cs
