+++ dotnet/src/libs-comparison.csproj
@@ -3,17 +3,17 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net8.0</TargetFramework>
-    <RootNamespace>libs_comparison</RootNamespace>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="BenchmarkDotNet" Version="0.13.5" />
+    <ProjectReference Include="..\deps\feetlicker\src\Feetlicker.csproj" />
   </ItemGroup>
 
   <ItemGroup>
-    <Content Include="../data.txt" CopyToOutputDirectory="Always" />
+    <Content Include="../../data.txt" CopyToOutputDirectory="Always" />
   </ItemGroup>
 
 </Project>
